lint, cmd/openlynt: add rule type for comments
This implements definition of "comment_group" rules - if a comment
matches a certain expression (regexp), then it must also match the
required expression.

In practice, this lets openlynt pick up comments containing "TODO",
which it can then check against a specific pattern and trigger a lint
error if it doesn't match.

Added a small section to the README for this configuration type.

fixes #1