lint, cmd/openlynt: filter violations by commit range
New optional configuration in openlynt.yml: `revisions`

```yaml
revisions:
  from: optional<string>
  to: optional<string>
```

If specified, violations are excluded from output if they are newer than
"from" and (optionally) older than "to".

If unspecified, no violations are filtered.

We _may_ want to still print the old violations, but not trigger a
failure. TBD.

The tool itself is probably due for an actual well-designed refactor at
this point, I'm still kinda hacking around with things...

fixes #2